+++ pkg/diff/compare.go
@@ -18,46 +18,58 @@ func Compare(config1, config2 bitesize.Environment) string {
 	c1.Name = ""
 	c2.Name = ""
 
-	// XXX: remove tprs for now
 	var newServices bitesize.Services
 	for _, s := range c1.Services {
-		// if s.Type == "" {
 		d := c2.Services.FindByName(s.Name)
 
-		if d != nil {
-			if s.Version == "" {
-				s.Version = d.Version
-			}
-			s.Status = d.Status
-		}
-		// Ignore diff for annotations if d.Deployment is nil
-		if d == nil || d.Deployment == nil {
-			s.Annotations = nil
-		}
+		alignServices(&s, d)
 
 		newServices = append(newServices, s)
-		// }
 	}
 	c1.Services = newServices
-
-	// Ignore diff between service replicas when hpa is configured
-	var correctedServices bitesize.Services
-	gitServices := c1.Services
-	clusterServices := c2.Services
-	for _, gs := range gitServices {
-		cs := clusterServices.FindByName(gs.Name)
-		if cs != nil {
-			if cs.HPA.MinReplicas != 0 {
-				gs.Replicas = cs.Replicas
-			}
-		}
-		correctedServices = append(correctedServices, gs)
-	}
-	c1.Services = correctedServices
 
 	compareConfig := &pretty.Config{
 		Diffable:       true,
 		SkipZeroFields: true,
 	}
 	return compareConfig.Compare(c2, c1)
+}
+
+// Can't think of a better word
+func alignServices(src, dest *bitesize.Service) {
+	if dest != nil {
+		// Copy version from dest if source version is empty
+		if src.Version == "" {
+			src.Version = dest.Version
+		}
+
+		// Copy status from dest (status is only stored in the cluster)
+		src.Status = dest.Status
+
+		// Override source replicas with dest replicas if HPA is active
+		if dest.HPA.MinReplicas != 0 {
+			src.Replicas = dest.Replicas
+		}
+
+		if dest.Version == "" {
+			// If no deployment yet, ignore annotations. They only apply onto
+			// deployment object.
+			src.Annotations = dest.Annotations
+		} else {
+			// Apply all existing annotations
+			for k, v := range dest.Annotations {
+				if src.Annotations[k] == "" {
+					src.Annotations[k] = v
+				}
+			}
+		}
+
+		// I don't like it. All of these should live somewhere else rather than in
+		// diff logic :(
+		if len(dest.DeployedPods) > 0 {
+			src.DeployedPods = dest.DeployedPods
+		}
+
+	}
+
 }