+++ CHANGELOG.md
@@ -13,11 +13,12 @@ This project adheres to [Semantic Versioning](http://semver.org/). Additionally,
 
 #### Added
 
+*  Support for kubernetes service annotations. [[BITE-1511](https://agile-jira.pearson.com/browse/BITE-1511)]
+
 ### **[0.0.3] - 2017-08-31 [RELEASED]**
 
 #### Added
 
-*  Support for kubernetes service annotations. [[BITE-1511](https://agile-jira.pearson.com/browse/BITE-1511)]
 *  Support for configuring horizontal pod autoscaling. [[BITE-1433](https://agile-jira.pearson.com/browse/BITE-1433)]
 *  Added new environment operator endpoint for Pod Status. [[BITE-1484](https://agile-jira.pearson.com/browse/BITE-1484)]
 *  Custom Docker registry support added for pod spec. [[BITE-1448](https://agile-jira.pearson.com/browse/BITE-1448)]